Moved editor code out into it's own folder.
