+++ code/ExampleCommands.cs
@@ -39,4 +39,43 @@ internal class ExampleCommands
             }
         }
     }
+
+    [ConCmd.Server("damageme", Help = "damage me")]
+    public static void DamageCommand(int amount = 10)
+    {
+        var client = ConsoleSystem.Caller;
+
+        if (client != null && client.Pawn is PlayerBase ply && ply.ActiveChild is WeaponBase weapon)
+        {
+            var damageInfo = new DamageInfo();
+            damageInfo.Damage = amount;
+            damageInfo.Weapon = weapon;
+
+            ply.TakeDamage(damageInfo);
+        }
+    }
+
+    [ConCmd.Server("swb_attachment_equip", Help = "Equips an attachment by name")]
+    public static void EquipAttachmentCMD(string name, bool enabled)
+    {
+        var client = ConsoleSystem.Caller;
+
+        if (client != null)
+        {
+            var player = client.Pawn as PlayerBase;
+
+            if (player.ActiveChild is not WeaponBase activeWeapon) return;
+
+            if (enabled && activeWeapon.GetAttachment(name) != null)
+            {
+                // Attach
+                _ = activeWeapon.EquipAttachmentSV(name);
+            }
+            else if (!enabled && activeWeapon.GetActiveAttachment(name) != null)
+            {
+                // Detach
+                activeWeapon.UnequipAttachmentSV(name);
+            }
+        }
+    }
 }
