+++ code/swb_editor/AttachmentEditor.cs
