+++ code/swb_editor/ModelEditor.cs
