+++ code/swb_editor/PlayerBase.Editor.cs
@@ -1,0 +1,12 @@
+﻿namespace SWB_Base;
+
+public partial class PlayerBase
+{
+    public bool IsLookingAtGun
+    {
+        get
+        {
+            return attachmentEditor != null || modelEditor != null;
+        }
+    }
+}
