+++ code/swb_editor/ui/AttachmentEditorMenu.cs
