+++ code/swb_editor/ui/AttachmentEditorMenu.scss
