+++ code/swb_editor/ui/ModelDisplay.cs
