+++ code/swb_editor/ui/ModelEditorMenu.cs
