+++ code/swb_editor/ui/components/Slider.scss
