+++ code/swb_editor/ui/components/SliderEntry.cs
