+++ code/swb_editor/ui/components/SliderEntry.scss
